requirements.txt
streamlit
fbprophet
pandas
pandas_datareader
Pillow
